feat: Added start of hub map and portals
Don't be alarmed by the gigantic diff. It is because I had to reimport the assests for the map from the map editor.

The game now loads into the hub map in the spawn area and is it not finished yet. Use portal entity with the loaded in reference to the world you would like it to teleport to for it to teleport you to it.